+++ src/scripts/get_task_dfn_val.py
@@ -2,14 +2,12 @@ from __future__ import absolute_import
 import sys
 import json
 
-
 def run(element_name, task_definition_str):
     try:
         definition = json.loads(task_definition_str)
         task_definition = definition['taskDefinition']
     except:
-        raise Exception('No valid task definition found: ' +
-                        task_definition_str)
+        raise Exception('No valid task definition found: ' + task_definition_str)
     str_list_types = ['requiresCompatibilities']
     json_arr_types = ['placementConstraints', 'volumes', 'tags']
     json_obj_types = ['proxyConfiguration']